// bridge/typechain-types/BToken.ts
export declare namespace BToken {
  export type DepositStruct = {
    accountType: BigNumberish;
    account: string;
    value: BigNumberish;
  };

  export type DepositStructOutput = [number, string, BigNumber] & {
    accountType: number;
    account: string;
    value: BigNumber;
  };
}